Simplify testing and expand test suites
